+++ tools/tbench/w11a_pcnt/test_pcnt_basics.tcl
@@ -1,16 +1,16 @@
-# $Id: test_pcnt_basics.tcl 1050 2018-09-23 15:46:42Z mueller $
+# $Id: test_pcnt_basics.tcl 1053 2018-10-06 20:34:52Z mueller $
 #
 # Copyright 2018- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 # License disclaimer see License.txt in $RETROBASE directory
 #
 # Revision History:
 # Date         Rev Version  Comment
-# 2018-09-23  1450   1.0    Initial version
+# 2018-09-23  1050   1.0    Initial version
 #
 # Test basic perf counter functionality 
 
 # ----------------------------------------------------------------------------
-rlc log "test_pcnt_regs: test register response ------------------------------"
+rlc log "test_pcnt_basics: test basic functionality --------------------------"
 
 if {[$cpu get haspcnt] == 0} {
   rlc log "  test_pcnt_regs-W: no pcnt unit found, test aborted"
@@ -20,13 +20,13 @@ if {[$cpu get haspcnt] == 0} {
 # -- Section A ---------------------------------------------------------------
 rlc log "  A: simple loop code ---------------------------------------"
 
-cpu0 ldasm -lst lst -sym sym {
+$cpu ldasm -lst lst -sym sym {
         . = 1000
 stack:  
-start:  clr     r0
-        mov     #32.,r1
-1$:     inc     r0
-        sob     r1,1$
+start:  clr     r1
+        mov     #32.,r0
+1$:     inc     r1
+        sob     r0,1$
         halt
 stop:   
 }
@@ -40,7 +40,7 @@ $cpu cp \
 # run code
 rw11::asmrun  $cpu sym
 rw11::asmwait $cpu sym 
-rw11::asmtreg $cpu r0 32 r1 0
+rw11::asmtreg $cpu r0 0 r1 32
 # stop pcnt
 $cpu cp \
   -wreg pc.cntl [regbld rw11::PC_CNTL {func "STO"}] \
@@ -52,11 +52,11 @@ $cpu cp \
 #     4      =0    cpu_sm
 #     5      =0    cpu_um
 #     6      67    cpu_inst
-#     7      =0    cpu_vfetch
-#     8      =0    cpu_irupt
-#     9      33    cpu_pcload
+#     7      31    cpu_pcload
+#     8      =0    cpu_vfetch
+#     9      =0    cpu_irupt
 rlc log "    A2: test random access (ainc=0) --------------------"
-# read pc(6) twice, (9) once, check status
+# read pc(6) twice, (9) once, (7) one, check status
 $cpu cp \
   -wreg pc.cntl [regbldkv rw11::PC_CNTL func "LOA" caddr 6 ainc 0] \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  6 ainc 0] \
@@ -70,10 +70,13 @@ $cpu cp \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  6 waddr 0 ainc 0] \
   -wreg pc.cntl [regbldkv rw11::PC_CNTL func "LOA" caddr 9 ainc 0] \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  9 ainc 0] \
-  -rreg pc.data -edata 33 \
-  -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  9 waddr 1 ainc 0] \
   -rreg pc.data -edata  0 \
-  -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  9 waddr 0 ainc 0] 
+  -wreg pc.cntl [regbldkv rw11::PC_CNTL func "LOA" caddr 7 ainc 0] \
+  -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  7 ainc 0] \
+  -rreg pc.data -edata 31 \
+  -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  7 waddr 1 ainc 0] \
+  -rreg pc.data -edata  0 \
+  -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  7 waddr 0 ainc 0] 
   
 rlc log "    A3: test sequential access (ainc=1) ----------------"
 # read pc(6) to pc(9) check status
@@ -84,11 +87,11 @@ $cpu cp \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  6 waddr 1 ainc 1] \
   -rreg pc.data -edata  0 \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  7 waddr 0 ainc 1] \
+  -rreg pc.data -edata 31 \
   -rreg pc.data -edata  0 \
   -rreg pc.data -edata  0 \
   -rreg pc.data -edata  0 \
   -rreg pc.data -edata  0 \
-  -rreg pc.data -edata 33 \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  9 waddr 1 ainc 1] \
   -rreg pc.data -edata  0 \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr 10 waddr 0 ainc 1]
@@ -98,7 +101,7 @@ rlc log "    A3: test block access (ainc=1) ---------------------"
 $cpu cp \
   -wreg pc.cntl [regbldkv rw11::PC_CNTL func "LOA" caddr  3 ainc 1] \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr  3 ainc 1] \
-  -rblk pc.data 14 -edata {0 0 0 0 0 0 67 0 0 0 0 0 33 0} \
+  -rblk pc.data 14 -edata {0 0  0 0  0 0  67 0  31 0  0 0  0 0} \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr 10 waddr 0 ainc 1]
 
 #rlc log "    A4: test clear -------------------------------------"
@@ -106,5 +109,5 @@ $cpu cp \
   -wreg pc.cntl [regbldkv rw11::PC_CNTL func "CLR"] \
   -wreg pc.cntl [regbldkv rw11::PC_CNTL func "LOA" caddr  3 ainc 1] \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr 3 ainc 1] \
-  -rblk pc.data 14 -edata {0 0 0 0 0 0  0 0 0 0 0 0  0 0} \
+  -rblk pc.data 14 -edata {0 0  0 0  0 0  0 0  0 0  0 0  0 0} \
   -rreg pc.stat -edata [regbldkv rw11::PC_STAT caddr 10 waddr 0 ainc 1]