+++ tools/tbench/w11a_pcnt/test_pcnt_regs.tcl
@@ -1,11 +1,12 @@
-# $Id: test_pcnt_regs.tcl 1050 2018-09-23 15:46:42Z mueller $
+# $Id: test_pcnt_regs.tcl 1053 2018-10-06 20:34:52Z mueller $
 #
 # Copyright 2018- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 # License disclaimer see License.txt in $RETROBASE directory
 #
 # Revision History:
 # Date         Rev Version  Comment
-# 2018-09-23  1450   1.0    Initial version
+# 2018-10-06  1053   1.0    Initial version
+# 2018-09-23  1050   0.1    First draft
 #
 # Test register response 
 
@@ -21,14 +22,15 @@ if {[$cpu get haspcnt] == 0} {
 rlc log "  A basic register access tests -----------------------------"
 
 rlc log "    A1: write cntl, read stat --------------------------"
-# test start,stop works and run flag follows
+# test start,stop works and run flag follows; test cntl readable (no check)
 $cpu cp \
   -wreg pc.cntl [regbld rw11::PC_CNTL {func "CLR"}] \
   -rreg pc.stat -edata [regbld rw11::PC_STAT] \
   -wreg pc.cntl [regbld rw11::PC_CNTL {func "STA"}] \
   -rreg pc.stat -edata [regbld rw11::PC_STAT run] \
   -wreg pc.cntl [regbld rw11::PC_CNTL {func "STO"}] \
-  -rreg pc.stat -edata [regbld rw11::PC_STAT]
+  -rreg pc.stat -edata [regbld rw11::PC_STAT] \
+  -rreg pc.cntl 
 # test that load works, caddr and ainc follow in status, and that clr clears
 $cpu cp \
   -wreg pc.cntl [regbldkv rw11::PC_CNTL func "LOA" caddr 0x07 ainc 0] \